requirements.txt
pytest
instagrapi
python-telegram-bot~=22.3
httpx~=0.28.1
dotenv~=0.9.9
python-dotenv~=1.1.1